Cæsar Ⅲ — Syntax update
* Update to script 2.11 syntax
* Improve virtual desktop trick to use current desktop resolution
* Complete dependencies list

+++ play.it-2/games/play-caesar-3.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -29,19 +29,17 @@ set -o errexit
 ###
 
 ###
-# Caesar III
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# Cæsar Ⅲ
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200321.1
 
 # Set game-specific variables
 
 GAME_ID='caesar-3'
-GAME_NAME='Caesar III'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
+GAME_NAME='Cæsar Ⅲ'
 
 ARCHIVE_GOG='setup_caesar3_2.0.0.9.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/caesar_3'
@@ -49,33 +47,29 @@ ARCHIVE_GOG_MD5='2ee16fab54493e1c2a69122fd2e56635'
 ARCHIVE_GOG_SIZE='550000'
 ARCHIVE_GOG_VERSION='1.1-gog2.0.0.9'
 
-ARCHIVE_DOC1_PATH='tmp'
-ARCHIVE_DOC1_FILES='./gog_eula.txt ./eula.txt'
-
-ARCHIVE_DOC2_PATH='app'
-ARCHIVE_DOC2_FILES='./readme.txt ./*.pdf'
+ARCHIVE_DOC_DATA_PATH='app'
+ARCHIVE_DOC_DATA_FILES='readme.txt *.pdf'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./*.dll ./*.exe ./*.inf ./*.ini'
+ARCHIVE_GAME_BIN_FILES='*.dll *.exe *.inf *.ini'
 
 ARCHIVE_GAME_MOVIES_PATH='app'
-ARCHIVE_GAME_MOVIES_FILES='./smk'
+ARCHIVE_GAME_MOVIES_FILES='smk'
 
 ARCHIVE_GAME_SOUNDS_PATH='app'
-ARCHIVE_GAME_SOUNDS_FILES='./wavs'
+ARCHIVE_GAME_SOUNDS_FILES='wavs'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./555 ./*.555 ./*.emp ./*.eng ./*.map ./*.sg2 ./c3_model.txt ./mission1.pak'
+ARCHIVE_GAME_DATA_FILES='555 *.555 *.emp *.eng *.map *.sg2 c3_model.txt mission1.pak'
 
 CONFIG_FILES='./caesar3.ini'
 DATA_FILES='./c3_model.txt ./status.txt ./*.sav'
 
-APP_WINETRICKS='vd=1024x768'
+APP_WINETRICKS="vd=\$(xrandr|awk '/\\*/ {print \$1}')"
 
 APP_MAIN_TYPE='wine'
-APP_MAIN_EXE='./c3.exe'
-APP_MAIN_ICON='./c3.exe'
-APP_MAIN_ICON_RES='16 32'
+APP_MAIN_EXE='c3.exe'
+APP_MAIN_ICON='c3.exe'
 
 PACKAGES_LIST='PKG_MOVIES PKG_SOUNDS PKG_DATA PKG_BIN'
 
@@ -87,57 +81,54 @@ PKG_SOUNDS_DESCRIPTION='sounds'
 
 PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
+PKG_DATA_DEPS="$PKG_MOVIES_ID $PKG_SOUNDS_ID"
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_MOVIES_ID, $PKG_SOUNDS_ID, $PKG_DATA_ID, wine32-development | wine32 | wine-bin | wine-i386 | wine-staging-i386, wine:amd64 | wine"
-PKG_BIN_DEPS_ARCH="$PKG_MOVIES_ID $PKG_SOUNDS_ID $PKG_DATA_ID wine"
+PKG_BIN_DEPS="$PKG_DATA_ID wine winetricks xrandr"
 
 # Load common functions
 
-target_version='2.0'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
-
-PKG='PKG_MOVIES'
-organize_data 'GAME_MOVIES' "$PATH_GAME"
-
-PKG='PKG_SOUNDS'
-organize_data 'GAME_SOUNDS' "$PATH_GAME"
-
-PKG='PKG_DATA'
-organize_data 'DOC1'      "$PATH_DOC"
-organize_data 'DOC2'      "$PATH_DOC"
-organize_data 'GAME_DATA' "$PATH_GAME"
-
-PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
+prepare_package_layout
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Get game icon
+
+PKG='PKG_BIN'
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
